requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
gunicorn>=20.0.4
# Additional dependencies for testing
pytest>=7.0.0
pytest-django>=4.5.0